refactor: remove DateRange files and its utils
BREAKING CHANGE: This commit deletes all files related to DateRange.
The code has been moved to a separate repository called calendar-range,
which can be accessed at https://github.com/jschepke/calendar-range

DateRange was never meant to be a part of the auth project.
It was a result of testing and prototyping that started to live
its own life.

The only right direction was to move it to its own codebase where
it can be developed and maintained independently.